requirements.txt
solana==0.35.1
solders==0.21.0
anchorpy==0.20.1
python-dotenv==1.0.0
base58==2.1.1
websockets==10.4  # Compatible with both solana and anchorpy
asyncio==3.4.3
aiohttp==3.9.1
numpy==1.26.2
pandas==2.1.3
python-binance==1.0.19  # For price data comparison
ccxt==4.1.13  # For multi-exchange support
web3==6.11.3  # For cross-chain monitoring
